requirements.txt
Flask
fpdf2
Pillow
gunicorn # Recommended for production deployment on Render